requirements.txt
fastapi==0.103.0
uvicorn== 0.20.0
torch== 2.4.0
pillow==10.4.0
torchvision==0.19.0
